Add files via upload
"Hello everyone, and welcome back to Mayuresh Code, we're going to create a fun and interactive Amazon quiz app using HTML, CSS, and JavaScript.

We'll start by designing the layout of our quiz app using HTML to structure the content. Then, we'll style it with CSS to make it visually appealing. Finally, we'll use JavaScript to add functionality such as displaying questions, tracking answers, and calculating scores.

This project is a great way to practice your web development skills and create something cool at the same time. So, let's dive in and build our Amazon quiz app from scratch!